Updated Styles and Renderer
Homebrewery has made some changes to the Renderer as well as the
styles as part of what I believe to be a big update to Homebrewery.

I've updated the parsers and files so that the legacy renderer
works while allowing me to add support for the new renderer when
ready.